+++ CHANGELOG.md
@@ -11,8 +11,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Fixed bug in argo workflow which allows to successfully run on INT-A/INT-B
 ### Changed
 - Fixed table-settings reset bug
+- Changed datepicker in FE to date range picker
 - Fixed name of veracode backend job
 - Bump jetty-http from 11.0.15 to 11.0.17
+
 ### Removed
 
 ## [9.0.0-rc1 - 03.11.2023]